+++ R/endpoints.R
@@ -17,19 +17,24 @@ api_build <- function(queue) {
   pr$handle("GET", "/meta/plotting/<iso3>", endpoint_plotting_metadata,
             serializer = serializer_json_hintr())
   pr$handle("GET", "/download/spectrum/<id>", endpoint_download_spectrum(queue),
-            serializer = serializer_zip())
+            serializer = serializer_zip("naomi_spectrum_digest"))
   pr$handle("HEAD", "/download/spectrum/<id>", endpoint_download_spectrum(queue),
-            serializer = serializer_zip())
+            serializer = serializer_zip("naomi_spectrum_digest"))
   pr$handle("GET", "/download/summary/<id>", endpoint_download_summary(queue),
-            serializer = serializer_zip())
+            serializer = serializer_zip("naomi_summary"))
   pr$handle("HEAD", "/download/summary/<id>", endpoint_download_summary(queue),
-            serializer = serializer_zip())
+            serializer = serializer_zip("naomi_summary"))
   pr$handle("GET", "/hintr/version", endpoint_hintr_version,
             serializer = serializer_json_hintr())
   pr$handle("GET", "/hintr/worker/status", endpoint_hintr_worker_status(queue),
             serializer = serializer_json_hintr())
   pr$handle("POST", "/hintr/stop", endpoint_hintr_stop(queue))
   pr$handle("GET", "/", api_root)
+
+  pr$registerHook("preroute", api_log_start)
+  pr$registerHook("postserialize", api_log_end)
+  pr$set404Handler(hintr_404)
+
   pr
 }
 
@@ -40,6 +45,25 @@ api_run <- function(pr, port = 8888) {
 api <- function(port = 8888, queue_id = NULL, workers = 2) {
   queue <- Queue$new(queue_id, workers) # nocov
   api_run(api_build(queue), port) # nocov
+}
+
+api_log_start <- function(data, req, res) {
+  api_log("%s %s", req$REQUEST_METHOD, req$PATH_INFO)
+}
+
+api_log_end <- function(data, req, res, value) {
+  if (is.raw(value$body)) {
+    size <- length(value$body)
+  } else {
+    size <- nchar(value$body)
+  }
+  api_log("`--> %d (%d bytes)", value$status, size)
+  value
+}
+
+# We can route this via some check for enabling/disabling logging later
+api_log <- function(fmt, ...) {
+  message(sprintf("[%s] %s", Sys.time(), sprintf(fmt, ...)))
 }
 
 #' Get function to generate model options from Naomi template and input files
@@ -283,7 +307,11 @@ download <- function(queue, type) {
     path <- switch(type,
                    "spectrum" = response$value$spectrum_path,
                    "summary" = response$value$summary_path)
-    readBin(path, "raw", n = file.size(path))
+    out <- list(
+      bytes = readBin(path, "raw", n = file.size(path)),
+      id = id
+    )
+    out
   }
 }
 
@@ -304,9 +332,13 @@ endpoint_plotting_metadata <- function(req, res, iso3) {
 #' @param value List containing an indication of success, any errors and the
 #' value to return.
 #'
+#' @param as_json Logical, indicating if the response should be
+#'   converted into a json string
+#'
 #' @return Formatted hintr response.
 #' @keywords internal
-hintr_response <- function(value, schema) {
+#' @noRd
+hintr_response <- function(value, schema, as_json = TRUE) {
   if (value$success) {
     status <- "success"
   } else {
@@ -318,15 +350,16 @@ hintr_response <- function(value, schema) {
   else {
     errors = value$errors
   }
-  ret <- to_json(list(
+  response <- list(
     status = scalar(status),
     errors = errors,
-    data = value$value))
+    data = value$value)
+  ret <- to_json(response)
   if (value$success) {
     validate_json_schema(ret, schema, query = "data")
   }
   validate_json_schema(ret, "Response")
-  ret
+  if (as_json) ret else response
 }
 
 hintr_errors <- function(errors) {
@@ -416,14 +449,33 @@ serializer_json_hintr <- function() {
   }
 }
 
-serializer_zip <- function() {
+serializer_zip <- function(filename) {
   function(val, req, res, errorHandler) {
     tryCatch({
       res$setHeader("Content-Type", "application/octet-stream")
-      res$body <- val
+      short_id <- substr(val$id, 1, 5)
+      res$setHeader("Content-Disposition",
+                    sprintf('attachment; filename="%s_%s.zip"',
+                            filename, short_id))
+      res$body <- val$bytes
       return(res$toResponse())
     }, error = function(e) {
       errorHandler(req, res, e)
     })
   }
+}
+
+hintr_404 <- function(req, res) {
+  res$status <- 404L
+  detail <- sprintf("%s %s is not a valid hintr path",
+                    req$REQUEST_METHOD, req$PATH_INFO)
+  errors <- hintr_errors(list("NOT_FOUND" = detail))
+  value <- list(success = FALSE,
+                errors = errors)
+  # We get no control over how plumber will serialise this - so we
+  # can't return json or it gets wrapped as if it was a json string
+  # (we ordinarily get around this by using json_verbatim = TRUE).  So
+  # here we return the object that will be passed into
+  # jsonlite::toJSON (all scalars being appropriately treated).
+  hintr_response(value, as_json = FALSE)
 }