+++ tests/testthat/test-server.R
@@ -430,6 +430,8 @@ test_that("spectrum file download streams bytes", {
     r <- httr::GET(paste0(server$url, "/download/spectrum/", response$data$id))
     expect_equal(httr::status_code(r), 200)
     expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+    expect_match(httr::headers(r)$`content-disposition`,
+                 'attachment; filename="naomi_spectrum_digest_\\w+.zip"')
     ## Size of bytes is close to expected
     size <- as.numeric(httr::headers(r)$`content-length`)
     expect_true(size - size/10 <
@@ -442,6 +444,8 @@ test_that("spectrum file download streams bytes", {
   r <- httr::HEAD(paste0(server$url, "/download/spectrum/", response$data$id))
   expect_equal(httr::status_code(r), 200)
   expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+  expect_match(httr::headers(r)$`content-disposition`,
+    'attachment; filename="naomi_spectrum_digest_\\w+.zip"')
 })
 
 test_that("summary file download streams bytes", {
@@ -465,6 +469,8 @@ test_that("summary file download streams bytes", {
     r <- httr::GET(paste0(server$url, "/download/summary/", response$data$id))
     expect_equal(httr::status_code(r), 200)
     expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+    expect_match(httr::headers(r)$`content-disposition`,
+                 'attachment; filename="naomi_summary_\\w+.zip"')
     ## Size of bytes is close to expected
     size <- as.numeric(httr::headers(r)$`content-length`)
     expect_true(size - size/10 <
@@ -477,6 +483,8 @@ test_that("summary file download streams bytes", {
   r <- httr::HEAD(paste0(server$url, "/download/summary/", response$data$id))
   expect_equal(httr::status_code(r), 200)
   expect_equal(httr::headers(r)$`content-type`, "application/octet-stream")
+  expect_match(httr::headers(r)$`content-disposition`,
+               'attachment; filename="naomi_summary_\\w+.zip"')
 })
 
 test_that("can quit", {
@@ -493,4 +501,18 @@ test_that("can quit", {
   expect_is(r, "error")
 
   expect_false(server$process$is_alive())
+})
+
+test_that("404 pages have sensible schema", {
+  server <- hintr_server()
+  r <- httr::GET(paste0(server$url, "/meaning-of-life"))
+  expect_equal(r$status_code, 404)
+  expect_equal(r$headers[["content-type"]], "application/json")
+
+  dat <- httr::content(r, "parsed", encoding = "UTF-8")
+  expect_equal(dat$status, "failure")
+  expect_equal(dat$errors[[1]]$error,
+               "NOT_FOUND")
+  expect_equal(dat$errors[[1]]$detail,
+               "GET /meaning-of-life is not a valid hintr path")
 })